test(maru): split MaruMultiValidatorTest into 2 parallel classes
MaruMultiValidatorTest ran its 4 heavy QBFT scenarios serially in a single
Gradle fork (forkEvery=1 + same_thread). On CI that fork is the makespan
tail: ~536s of one mostly-idle worker (each scenario spins up 4 in-process
validators and waits for convergence with 240s timeouts), while the other
~7 cores sit unused.

Extract MaruMultiValidatorTestBase and split the scenarios into
MaruMultiValidatorConvergenceTest (convergence/liveness) and
MaruMultiValidatorRecoveryTest (restart recovery) so two run in parallel
forks (~halving the tail). Deliberately a 2-way split, not one class per
scenario: four heavy consensus forks on 1s slot timers can starve those
timers under CFS throttling and push convergence toward the 240s timeout.
