+++ maru/app/src/integrationTest/kotlin/maru/app/MaruMultiValidatorTestBase.kt
@@ -27,7 +27,6 @@ import org.hyperledger.besu.tests.acceptance.dsl.node.cluster.ClusterConfigurati
 import org.hyperledger.besu.tests.acceptance.dsl.transaction.net.NetTransactions
 import org.junit.jupiter.api.AfterEach
 import org.junit.jupiter.api.BeforeEach
-import org.junit.jupiter.api.Test
 import testutils.PeeringNodeNetworkStack
 import testutils.maru.MaruFactory
 import testutils.maru.awaitTillMaruHasPeers
@@ -36,33 +35,43 @@ import kotlin.time.Duration.Companion.milliseconds
 import kotlin.time.Duration.Companion.seconds
 import kotlin.time.toJavaDuration
 
-class MaruMultiValidatorTest {
+/**
+ * Shared 4-validator QBFT fixture + helpers. The scenarios are split across concrete subclasses
+ * (convergence/liveness vs. restart-recovery) so they run in parallel Gradle forks instead of
+ * serially in one ~9-min fork. Each scenario still builds its own independent cluster in
+ * [setUp]/[startAllValidators], so the split is purely fork-level parallelism, not shared state.
+ *
+ * Deliberately kept to a 2-way split rather than one-class-per-scenario: each scenario runs four
+ * in-process validators on 1 s slot timers, and too many heavy consensus forks in parallel can
+ * starve those timers under CFS throttling and push convergence toward the 240 s timeout.
+ */
+abstract class MaruMultiValidatorTestBase {
   companion object {
     /** Number of consecutive round-0 blocks required to declare convergence / stable production. */
-    private const val STABLE_BLOCKS = 5
-
-    private val multiValidatorSyncingConfig = MaruFactory.defaultValidatorSyncingConfig
+    const val STABLE_BLOCKS = 5
   }
+
+  private val multiValidatorSyncingConfig = MaruFactory.defaultValidatorSyncingConfig
 
   private val key0 = "080212201dd171cec7e2995408b5513004e8207fe88d6820aeff0d82463b3e41df251aae".fromHex()
   private val key1 = "0802122100abb81ba53518eb0a206dfe80f2a973182e5d66c98cd31d00bf7471fcd5514157".fromHex()
   private val key2 = "080212202fec0750fe3edc7e8272d4814a36b632921fc5e835d20a2de874471e8ad9ad0b".fromHex()
   private val key3 = "080212207a19c01ce2246b94b48ed778d9bfb3b76eaabe8193c468d6751f4e4d1adf98a8".fromHex()
 
-  private lateinit var cluster: Cluster
-  private lateinit var stack0: PeeringNodeNetworkStack
-  private lateinit var stack1: PeeringNodeNetworkStack
-  private lateinit var stack2: PeeringNodeNetworkStack
-  private lateinit var stack3: PeeringNodeNetworkStack
+  protected lateinit var cluster: Cluster
+  protected lateinit var stack0: PeeringNodeNetworkStack
+  protected lateinit var stack1: PeeringNodeNetworkStack
+  protected lateinit var stack2: PeeringNodeNetworkStack
+  protected lateinit var stack3: PeeringNodeNetworkStack
 
-  private val log = LogManager.getLogger(this.javaClass)
+  protected val log = LogManager.getLogger(this.javaClass)
 
-  private val maruFactory0 = MaruFactory(validatorPrivateKey = key0)
-  private val maruFactory1 = MaruFactory(validatorPrivateKey = key1)
-  private val maruFactory2 = MaruFactory(validatorPrivateKey = key2)
-  private val maruFactory3 = MaruFactory(validatorPrivateKey = key3)
+  protected val maruFactory0 = MaruFactory(validatorPrivateKey = key0)
+  protected val maruFactory1 = MaruFactory(validatorPrivateKey = key1)
+  protected val maruFactory2 = MaruFactory(validatorPrivateKey = key2)
+  protected val maruFactory3 = MaruFactory(validatorPrivateKey = key3)
 
-  private val initialValidators: Set<Validator> by lazy {
+  protected val initialValidators: Set<Validator> by lazy {
     setOf(
       SecpCrypto.privateKeyToValidator(SecpCrypto.privateKeyBytesWithoutPrefix(key0)),
       SecpCrypto.privateKeyToValidator(SecpCrypto.privateKeyBytesWithoutPrefix(key1)),
@@ -101,7 +110,7 @@ class MaruMultiValidatorTest {
 
   // -- Helper methods ---------------------------------------------------------
 
-  private fun startAllValidators() {
+  protected fun startAllValidators() {
     val app0 =
       maruFactory0.buildTestMaruValidatorWithP2pPeering(
         ethereumJsonRpcUrl = stack0.besuNode.jsonRpcBaseUrl().get(),
@@ -179,7 +188,7 @@ class MaruMultiValidatorTest {
     log.info("All 4 validators peered in full mesh")
   }
 
-  private fun waitForBlockHeight(
+  protected fun waitForBlockHeight(
     vararg beaconChains: BeaconChain,
     targetHeight: ULong,
     timeout: Duration = 240.seconds,
@@ -204,7 +213,7 @@ class MaruMultiValidatorTest {
    * before the P2P mesh is wired, causing round skips. Checking a fixed block number is unreliable;
    * instead we wait for a stable run of round-0 blocks.
    */
-  private fun waitForConsecutiveRound0Blocks(
+  protected fun waitForConsecutiveRound0Blocks(
     beaconChain: BeaconChain,
     requiredConsecutive: Int = 5,
     timeout: Duration = 240.seconds,
@@ -242,17 +251,17 @@ class MaruMultiValidatorTest {
     return lastStableBlock
   }
 
-  private fun currentBlockHeight(stack: PeeringNodeNetworkStack): ULong =
+  protected fun currentBlockHeight(stack: PeeringNodeNetworkStack): ULong =
     stack.maruApp.beaconChain
       .getLatestBeaconState()
       .beaconBlockHeader.number
 
-  private fun stopValidator(stack: PeeringNodeNetworkStack) {
+  protected fun stopValidator(stack: PeeringNodeNetworkStack) {
     stack.maruApp.stop().get()
     stack.maruApp.close()
   }
 
-  private fun restartValidator(
+  protected fun restartValidator(
     stack: PeeringNodeNetworkStack,
     factory: MaruFactory,
     peersToConnect: List<MaruApp>,
@@ -276,7 +285,7 @@ class MaruMultiValidatorTest {
     app.awaitTillMaruHasPeers(peersToConnect.size.toUInt(), pollingInterval = 500.milliseconds)
   }
 
-  private fun <T, M> checkAllValidatorBlocksAreTheSame(
+  protected fun <T, M> checkAllValidatorBlocksAreTheSame(
     validatorBlocks: List<() -> List<T>>,
     blocksToMetadata: (List<T>) -> List<M>,
   ) {
@@ -288,7 +297,7 @@ class MaruMultiValidatorTest {
     }
   }
 
-  private fun checkBlockProposersMatchExpectedProposers(
+  protected fun checkBlockProposersMatchExpectedProposers(
     beaconChain: BeaconChain,
     startBlock: ULong,
     endBlock: ULong,
@@ -312,208 +321,10 @@ class MaruMultiValidatorTest {
     }
   }
 
-  private fun clBlocksToMetadata(blocks: List<SealedBeaconBlock>): List<Pair<ULong, String>> =
+  protected fun clBlocksToMetadata(blocks: List<SealedBeaconBlock>): List<Pair<ULong, String>> =
     blocks.map {
       it.beaconBlock.beaconBlockHeader.number to
         it.beaconBlock.beaconBlockHeader.hash
           .encodeHex()
     }
-
-  // -- Test scenarios ---------------------------------------------------------
-
-  @Test
-  fun `validators converge to stable block production without round skips`() {
-    startAllValidators()
-
-    // Wait until 5 consecutive round-0 blocks are observed. During startup, validators run QBFT
-    // independently before the P2P mesh is wired, causing round skips on early blocks
-    val stableHeight =
-      waitForConsecutiveRound0Blocks(
-        stack0.maruApp.beaconChain,
-        requiredConsecutive = STABLE_BLOCKS,
-        timeout = 240.seconds,
-      )
-    log.info("QBFT convergence achieved at block $stableHeight")
-
-    // Verify STABLE_BLOCKS more blocks after convergence are also round-0.
-    // Wait for ALL validators to reach the target so blocks can safely be read from all.
-    waitForBlockHeight(
-      stack0.maruApp.beaconChain,
-      stack1.maruApp.beaconChain,
-      stack2.maruApp.beaconChain,
-      stack3.maruApp.beaconChain,
-      targetHeight = stableHeight + STABLE_BLOCKS.toULong(),
-      timeout = 90.seconds,
-    )
-    val verifyStart = stableHeight - (STABLE_BLOCKS - 1).toULong()
-    val verifyCount = (STABLE_BLOCKS * 2).toULong()
-    val verifyBlocks = stack0.maruApp.beaconChain.getSealedBeaconBlocks(verifyStart, verifyCount)
-    verifyBlocks.forEach { block ->
-      val header = block.beaconBlock.beaconBlockHeader
-      assertThat(header.round)
-        .withFailMessage { "Block ${header.number} has round ${header.round}, expected 0" }
-        .isEqualTo(0u)
-    }
-
-    // Verify blocks consistent across all 4 validators
-    checkAllValidatorBlocksAreTheSame(
-      validatorBlocks = listOf(
-        { stack0.maruApp.beaconChain.getSealedBeaconBlocks(verifyStart, verifyCount) },
-        { stack1.maruApp.beaconChain.getSealedBeaconBlocks(verifyStart, verifyCount) },
-        { stack2.maruApp.beaconChain.getSealedBeaconBlocks(verifyStart, verifyCount) },
-        { stack3.maruApp.beaconChain.getSealedBeaconBlocks(verifyStart, verifyCount) },
-      ),
-      blocksToMetadata = ::clBlocksToMetadata,
-    )
-
-    // Verify proposers match expected
-    checkBlockProposersMatchExpectedProposers(
-      beaconChain = stack0.maruApp.beaconChain,
-      startBlock = verifyStart,
-      endBlock = stableHeight + STABLE_BLOCKS.toULong(),
-    )
-  }
-
-  @Test
-  fun `block production continues with 1 node offline`() {
-    startAllValidators()
-
-    // Wait for convergence before stopping a node
-    waitForConsecutiveRound0Blocks(
-      stack0.maruApp.beaconChain,
-      requiredConsecutive = STABLE_BLOCKS,
-      timeout = 240.seconds,
-    )
-
-    // Stop validator 3
-    log.info("Stopping validator 3")
-    stopValidator(stack3)
-
-    // Record current height and wait for STABLE_BLOCKS more blocks
-    val heightAfterStop = currentBlockHeight(stack0)
-    log.info("Height after stopping validator 3: $heightAfterStop")
-    // Wait for all 3 remaining validators so getSealedBeaconBlocks doesn't throw on any of them.
-    waitForBlockHeight(
-      stack0.maruApp.beaconChain,
-      stack1.maruApp.beaconChain,
-      stack2.maruApp.beaconChain,
-      targetHeight = heightAfterStop + STABLE_BLOCKS.toULong(),
-      timeout = 240.seconds,
-    )
-
-    // Verify blocks consistent across the 3 remaining validators
-    val verifyStart = heightAfterStop + 1uL
-    val count = STABLE_BLOCKS.toULong()
-    checkAllValidatorBlocksAreTheSame(
-      validatorBlocks = listOf(
-        { stack0.maruApp.beaconChain.getSealedBeaconBlocks(verifyStart, count) },
-        { stack1.maruApp.beaconChain.getSealedBeaconBlocks(verifyStart, count) },
-        { stack2.maruApp.beaconChain.getSealedBeaconBlocks(verifyStart, count) },
-      ),
-      blocksToMetadata = ::clBlocksToMetadata,
-    )
-    log.info("Block production continued successfully with 3 of 4 validators")
-  }
-
-  @Test
-  fun `block production recovers after 2 nodes offline and 1 returns`() {
-    startAllValidators()
-
-    // Wait for convergence before stopping nodes
-    waitForConsecutiveRound0Blocks(
-      stack0.maruApp.beaconChain,
-      requiredConsecutive = STABLE_BLOCKS,
-      timeout = 240.seconds,
-    )
-
-    // Stop validators 2 and 3 -- only 2 of 4 remain, below quorum (need 3)
-    log.info("Stopping validators 2 and 3")
-    stopValidator(stack2)
-    stopValidator(stack3)
-
-    val heightAfterStop = currentBlockHeight(stack0)
-    log.info("Height after stopping 2 validators: $heightAfterStop")
-
-    // Wait 5 seconds and verify no new blocks were produced
-    Thread.sleep(5000)
-    val heightAfterWait = currentBlockHeight(stack0)
-    assertThat(heightAfterWait)
-      .withFailMessage {
-        "Expected no new blocks (height $heightAfterStop) but got height $heightAfterWait"
-      }.isEqualTo(heightAfterStop)
-    log.info("Confirmed: no blocks produced without quorum")
-
-    // Restart validator 2 -- quorum restored (3 of 4)
-    log.info("Restarting validator 2")
-    restartValidator(stack2, maruFactory2, listOf(stack0.maruApp, stack1.maruApp))
-
-    // Wait for all 3 active validators before reading their blocks.
-    waitForBlockHeight(
-      stack0.maruApp.beaconChain,
-      stack1.maruApp.beaconChain,
-      stack2.maruApp.beaconChain,
-      targetHeight = heightAfterStop + STABLE_BLOCKS.toULong(),
-      timeout = 90.seconds,
-    )
-
-    // Verify consistency across the 3 active validators
-    val verifyStart = heightAfterStop + 1uL
-    val count = STABLE_BLOCKS.toULong()
-    checkAllValidatorBlocksAreTheSame(
-      validatorBlocks = listOf(
-        { stack0.maruApp.beaconChain.getSealedBeaconBlocks(verifyStart, count) },
-        { stack1.maruApp.beaconChain.getSealedBeaconBlocks(verifyStart, count) },
-        { stack2.maruApp.beaconChain.getSealedBeaconBlocks(verifyStart, count) },
-      ),
-      blocksToMetadata = ::clBlocksToMetadata,
-    )
-    log.info("Block production recovered after quorum was restored")
-  }
-
-  @Test
-  fun `block production resumes after all 4 nodes restart`() {
-    startAllValidators()
-
-    // Wait for STABLE_BLOCKS blocks before recording the checkpoint (stack0 only; all are synced).
-    waitForBlockHeight(stack0.maruApp.beaconChain, targetHeight = STABLE_BLOCKS.toULong(), timeout = 90.seconds)
-    val heightBeforeRestart = currentBlockHeight(stack0)
-    log.info("Height before full restart: $heightBeforeRestart")
-
-    // Stop all 4 validators
-    log.info("Stopping all 4 validators")
-    stopValidator(stack0)
-    stopValidator(stack1)
-    stopValidator(stack2)
-    stopValidator(stack3)
-
-    Thread.sleep(2000)
-
-    // Restart all 4 and re-establish full mesh
-    log.info("Restarting all 4 validators")
-    startAllValidators()
-
-    // Wait for ALL 4 validators to reach the target before reading their blocks.
-    waitForBlockHeight(
-      stack0.maruApp.beaconChain,
-      stack1.maruApp.beaconChain,
-      stack2.maruApp.beaconChain,
-      stack3.maruApp.beaconChain,
-      targetHeight = heightBeforeRestart + STABLE_BLOCKS.toULong(),
-      timeout = 90.seconds,
-    )
-
-    // Verify consistency across all 4 validators
-    val verifyStart = heightBeforeRestart + 1uL
-    val count = STABLE_BLOCKS.toULong()
-    checkAllValidatorBlocksAreTheSame(
-      validatorBlocks = listOf(
-        { stack0.maruApp.beaconChain.getSealedBeaconBlocks(verifyStart, count) },
-        { stack1.maruApp.beaconChain.getSealedBeaconBlocks(verifyStart, count) },
-        { stack2.maruApp.beaconChain.getSealedBeaconBlocks(verifyStart, count) },
-        { stack3.maruApp.beaconChain.getSealedBeaconBlocks(verifyStart, count) },
-      ),
-      blocksToMetadata = ::clBlocksToMetadata,
-    )
-    log.info("Block production resumed successfully after full restart")
-  }
 }
